First commit.
When running this script, you need to add command arguments for paths
for source directory and for destination directory. Script then checks
whether the destination dir already exists. If it doesn't then script
would copy the whole source tree. If it does exists then the script
goes through every file or dir in source dir and check if it is
also in destination dir. If it isn't then it is copied there.