Logging of changes now works properly.
When running the script user has to provide path to log_file as
third commandline argument (after source directory and destination
directory). When files or directories are copied to dest dir or
they are removed from there, the script logs this change in given
log_file and also in console.